load visible textures only
Pinpoint (or the underlying clutter library) does not cope well with a large
amount of texture data; either the program will crash if too many background
images are used or background textures are lost due to lack of texture memory.

This change enables pinpoint to only load those images that are visible or
might be visible soon, such as the backgrounds used by the current and adjacent
slides.

All other textures a replaced by a 1x1 placeholder, freeing much needed texture
memory.